refactor(orchestra): relocate second driver-passthrough batch into LegacyExecutionBackend
Relocates tapOnPoint, tapOnPointV2, scrollVertical, setPermissions,
waitForAnimationToEnd, setLocation, setOrientation, addMedia, setAirplaneMode,
toggleAirplaneMode, setDarkMode, toggleDarkMode, assertDarkMode (expected=true),
and travel verbatim from Orchestra into LegacyExecutionBackend, following the
pattern established in Tasks 1.2-1.4. assertDarkMode(expected: Boolean) is kept
duplicated in Orchestra since AssertLightModeCommand still calls it with
expected=false. Extends RouterSeamTest to cover all 14 newly-routed command
types and swaps its "not yet relocated" example from ScrollCommand (now
relocated) to PasteTextCommand.
